Rename a few packages for consistency.
