+++ src/maxpower/network/tcp/framer/TCPFramerConstants.java
@@ -6,12 +6,12 @@
  *                                                                   *
  *********************************************************************/
 
-package maxpower.network.tcp.manyconn.framer;
+package maxpower.network.tcp.framer;
 
 import java.util.List;
 
-import maxpower.network.tcp.manyconn.framer.TCPFramerSM.FramerStates;
-import maxpower.network.tcp.manyconn.framer.proto.FramerProtocolSpec;
+import maxpower.network.tcp.framer.TCPFramerSM.FramerStates;
+import maxpower.network.tcp.framer.proto.FramerProtocolSpec;
 
 import com.maxeler.maxcompiler.v2.managers.DFEManager;
 import com.maxeler.maxcompiler.v2.utils.MathUtils;
