+++ src/maxpower/network/tcp/framer/proto/FIXProtoSpec.maxj
@@ -6,11 +6,11 @@
  *                                                                   *
  *********************************************************************/
 
-package maxpower.network.tcp.manyconn.framer.proto;
+package maxpower.network.tcp.framer.proto;
 
-import maxpower.network.tcp.manyconn.framer.TCPFramerSM;
-import maxpower.network.tcp.manyconn.framer.TCPFramerSM.FramerStates;
-import maxpower.network.tcp.manyconn.framer.TCPFramerTypes;
+import maxpower.network.tcp.framer.TCPFramerSM;
+import maxpower.network.tcp.framer.TCPFramerTypes;
+import maxpower.network.tcp.framer.TCPFramerSM.FramerStates;
 
 import com.maxeler.maxcompiler.v2.managers.DFEManager;
 import com.maxeler.maxcompiler.v2.statemachine.DFEsmAssignableValue;
