+++ test/maxpower/lmem/cpuaccess/LMemCpuAccessTest.maxj
@@ -1,4 +1,6 @@
-package maxpower.lmem.cpu_access;
+package maxpower.lmem.cpuaccess;
+
+import maxpower.lmem.cpuaccess.LMemCpuAccess;
 
 import com.maxeler.maxcompiler.v2.build.EngineParameters;
 import com.maxeler.maxcompiler.v2.managers.DFEModel;
